+++ client/entity/function_test.go
@@ -1,0 +1,48 @@
+// Licensed to the LF AI & Data foundation under one
+// or more contributor license agreements. See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership. The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License. You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package entity
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestFunctionSchema(t *testing.T) {
+	functions := []*Function{
+		NewFunction().WithName("text_bm25_emb").WithType(FunctionTypeBM25).WithInputFields("a", "b").WithOutputFields("c").WithParam("key", "value"),
+		NewFunction().WithName("other_emb").WithType(FunctionTypeTextEmbedding).WithInputFields("c").WithOutputFields("b", "a"),
+	}
+
+	for _, function := range functions {
+		funcSchema := function.ProtoMessage()
+		assert.Equal(t, function.Name, funcSchema.Name)
+		assert.Equal(t, function.Type, funcSchema.Type)
+		assert.Equal(t, function.InputFieldNames, funcSchema.InputFieldNames)
+		assert.Equal(t, function.OutputFieldNames, funcSchema.OutputFieldNames)
+		assert.Equal(t, function.Params, KvPairsMap(funcSchema.GetParams()))
+
+		nf := NewFunction()
+		nf.ReadProto(funcSchema)
+
+		assert.Equal(t, function.Name, nf.Name)
+		assert.Equal(t, function.Type, nf.Type)
+		assert.Equal(t, function.InputFieldNames, nf.InputFieldNames)
+		assert.Equal(t, function.OutputFieldNames, nf.OutputFieldNames)
+		assert.Equal(t, function.Params, nf.Params)
+	}
+}
