+++ client/entity/schema_test.go
@@ -37,56 +37,6 @@ func TestCL_CommonCL(t *testing.T) {
 	}
 }
 
-func TestFieldSchema(t *testing.T) {
-	fields := []*Field{
-		NewField().WithName("int_field").WithDataType(FieldTypeInt64).WithIsAutoID(true).WithIsPrimaryKey(true).WithDescription("int_field desc"),
-		NewField().WithName("string_field").WithDataType(FieldTypeString).WithIsAutoID(false).WithIsPrimaryKey(true).WithIsDynamic(false).WithTypeParams("max_len", "32").WithDescription("string_field desc"),
-		NewField().WithName("partition_key").WithDataType(FieldTypeInt32).WithIsPartitionKey(true),
-		NewField().WithName("array_field").WithDataType(FieldTypeArray).WithElementType(FieldTypeBool).WithMaxCapacity(128),
-		NewField().WithName("clustering_key").WithDataType(FieldTypeInt32).WithIsClusteringKey(true),
-		/*
-			NewField().WithName("default_value_bool").WithDataType(FieldTypeBool).WithDefaultValueBool(true),
-			NewField().WithName("default_value_int").WithDataType(FieldTypeInt32).WithDefaultValueInt(1),
-			NewField().WithName("default_value_long").WithDataType(FieldTypeInt64).WithDefaultValueLong(1),
-			NewField().WithName("default_value_float").WithDataType(FieldTypeFloat).WithDefaultValueFloat(1),
-			NewField().WithName("default_value_double").WithDataType(FieldTypeDouble).WithDefaultValueDouble(1),
-			NewField().WithName("default_value_string").WithDataType(FieldTypeString).WithDefaultValueString("a"),*/
-	}
-
-	for _, field := range fields {
-		fieldSchema := field.ProtoMessage()
-		assert.Equal(t, field.ID, fieldSchema.GetFieldID())
-		assert.Equal(t, field.Name, fieldSchema.GetName())
-		assert.EqualValues(t, field.DataType, fieldSchema.GetDataType())
-		assert.Equal(t, field.AutoID, fieldSchema.GetAutoID())
-		assert.Equal(t, field.PrimaryKey, fieldSchema.GetIsPrimaryKey())
-		assert.Equal(t, field.IsPartitionKey, fieldSchema.GetIsPartitionKey())
-		assert.Equal(t, field.IsClusteringKey, fieldSchema.GetIsClusteringKey())
-		assert.Equal(t, field.IsDynamic, fieldSchema.GetIsDynamic())
-		assert.Equal(t, field.Description, fieldSchema.GetDescription())
-		assert.Equal(t, field.TypeParams, KvPairsMap(fieldSchema.GetTypeParams()))
-		assert.EqualValues(t, field.ElementType, fieldSchema.GetElementType())
-		// marshal & unmarshal, still equals
-		nf := &Field{}
-		nf = nf.ReadProto(fieldSchema)
-		assert.Equal(t, field.ID, nf.ID)
-		assert.Equal(t, field.Name, nf.Name)
-		assert.EqualValues(t, field.DataType, nf.DataType)
-		assert.Equal(t, field.AutoID, nf.AutoID)
-		assert.Equal(t, field.PrimaryKey, nf.PrimaryKey)
-		assert.Equal(t, field.Description, nf.Description)
-		assert.Equal(t, field.IsDynamic, nf.IsDynamic)
-		assert.Equal(t, field.IsPartitionKey, nf.IsPartitionKey)
-		assert.Equal(t, field.IsClusteringKey, nf.IsClusteringKey)
-		assert.EqualValues(t, field.TypeParams, nf.TypeParams)
-		assert.EqualValues(t, field.ElementType, nf.ElementType)
-	}
-
-	assert.NotPanics(t, func() {
-		(&Field{}).WithTypeParams("a", "b")
-	})
-}
-
 type SchemaSuite struct {
 	suite.Suite
 }
@@ -101,7 +51,8 @@ func (s *SchemaSuite) TestBasic() {
 			"test_collection",
 			NewSchema().WithName("test_collection_1").WithDescription("test_collection_1 desc").WithAutoID(false).
 				WithField(NewField().WithName("ID").WithDataType(FieldTypeInt64).WithIsPrimaryKey(true)).
-				WithField(NewField().WithName("vector").WithDataType(FieldTypeFloatVector).WithDim(128)),
+				WithField(NewField().WithName("vector").WithDataType(FieldTypeFloatVector).WithDim(128)).
+				WithFunction(NewFunction()),
 			"ID",
 		},
 		{
@@ -122,6 +73,7 @@ func (s *SchemaSuite) TestBasic() {
 			s.Equal(sch.Description, p.GetDescription())
 			s.Equal(sch.EnableDynamicField, p.GetEnableDynamicField())
 			s.Equal(len(sch.Fields), len(p.GetFields()))
+			s.Equal(len(sch.Functions), len(p.GetFunctions()))
 
 			nsch := &Schema{}
 			nsch = nsch.ReadProto(p)
@@ -130,6 +82,7 @@ func (s *SchemaSuite) TestBasic() {
 			s.Equal(sch.Description, nsch.Description)
 			s.Equal(sch.EnableDynamicField, nsch.EnableDynamicField)
 			s.Equal(len(sch.Fields), len(nsch.Fields))
+			s.Equal(len(sch.Functions), len(nsch.Functions))
 			s.Equal(c.pkName, sch.PKFieldName())
 			s.Equal(c.pkName, nsch.PKFieldName())
 		})
